Move ResourceInfo under shared folder (refs #192)
