+++ app/pages/resource/_resource-page.scss
@@ -1,5 +1,3 @@
-@import 'info/resource-info';
-
 .resource-page {
   .help-text {
     color: $gray;
