+++ app/shared/_shared.scss
@@ -11,4 +11,5 @@
 @import './navbar/navbar';
 @import './resource-info-button/resource-info-button';
 @import './reservation-form/reservation-form';
+@import './resource-info/resource-info';
 @import './wrapped-text/wrapped-text';
