+++ app/shared/resource-info/ResourceInfo.js
