+++ app/shared/resource-info/ResourceInfo.spec.js
