+++ app/shared/resource-info/_resource-info.scss
