+++ app/shared/resource-info/index.js
